backend/requirements.txt
fastapi
uvicorn
python-multipart
openai
pypdf
python-dotenv
requests
edge-tts
google-generativeai
